moves ts-runner
