+++ examples/fruits-pipeline-worker/src/index.ts
@@ -1,5 +1,5 @@
 import { Elm } from "./Main.elm";
-import * as Tasks from "../../../src/runner";
+import * as Tasks from "../../../src-ts";
 import * as S3 from "./Aws/s3";
 import * as SQS from "./Aws/sqs";
 import * as SNS from "./Aws/sns";
