+++ tsconfig.json
@@ -5,13 +5,13 @@
     "target": "ESNext",
 
     // Module resolution
-    "baseUrl": "./src/runner",
+    "baseUrl": "src-ts",
     "esModuleInterop": true,
     "moduleResolution": "node",
 
     // Source Map
     "sourceMap": true,
-    "sourceRoot": "./src/runner",
+    "sourceRoot": "./src-ts",
 
     // Strict Checks
     "alwaysStrict": true,
@@ -25,7 +25,7 @@
     "noUnusedLocals": true,
     "noUnusedParameters": true
   },
-  "include": ["./src/runner/**/*.ts"],
+  "include": ["src-ts/**/*.ts"],
   "exclude": [
     "node_modules/**/*"
   ]
